move miner to its own package
