+++ miner/miner_test.go
@@ -1,0 +1,21 @@
+package miner
+
+import (
+	"testing"
+)
+
+func TestMine(t *testing.T) {
+
+}
+
+func TestSetMiningHeader(t *testing.T) {
+
+}
+
+func TestVerifyProofOfWork(t *testing.T) {
+
+}
+
+func TestVerifyMiningHeader(t *testing.T) {
+
+}
